requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
backports-zoneinfo==0.2.1
    # via
    #   -r requirements/base.txt
    #   django
click==8.1.7
    # via code-annotations
code-annotations==1.7.0
    # via -r requirements/test.in
coverage[toml]==7.4.4
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
exceptiongroup==1.2.0
    # via pytest
iniconfig==2.0.0
    # via pytest
jinja2==3.1.3
    # via code-annotations
markupsafe==2.1.5
    # via jinja2
openedx-atlas==0.6.0
    # via -r requirements/base.txt
packaging==24.0
    # via pytest
pbr==6.0.0
    # via stevedore
pluggy==1.4.0
    # via pytest
pytest==8.1.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==5.0.0
    # via -r requirements/test.in
pytest-django==4.8.0
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.1
    # via code-annotations
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.2.0
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.10.0
    # via
    #   -r requirements/base.txt
    #   asgiref
